Add support for room/bathroom doors
This adds support to specify the doors to the room.  As long as that door
is closed the light will be blocked from turning off.  Closing the door
is also treated as a motion event and will cause the light to turn on.

You can also override the default delay to have the light turn off
quicker after the door is opened by specifying 'door_open_delay' otherwise
it will turn off after the normal delay

Additional config options:
  door:
    - binary_sensor.bathroom_door
  door_open_delay: 5